swarm/storage/script: rename package to vm
