+++ swarm/storage/script/vm/reference_test.go
@@ -1,4 +1,4 @@
-package txscript
+package vm
 
 import (
 	"encoding/hex"
