+++ swarm/storage/script/vm/standard.go
@@ -1,4 +1,4 @@
-package txscript
+package vm
 
 import "fmt"
